Move MakeEmptyHeader to chain_makers
